requirements.txt
# ========== LangChain 核心 ==========
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# LLM 支持
openai>=1.52.0

# ========== CTF 密码学工具 ==========
pwntools>=4.12.0
pycryptodome>=3.20.0
cryptography>=42.0.0
sympy>=1.12
gmpy2>=2.1.5

# ========== 数据处理 ==========
pandas>=2.0.0
numpy>=1.24.0

# ========== 工具 ==========
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0

# ========== 测试 ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0
